Remove repetition in sidebar labels
In some sidebar sections in which multiple entries unnecessarily repeat
words, remove the repetitive words to make the sidebar easier to read.

As part of this work, add docs subsections for pages that we currently
demarcate from the rest of a subsection using repeated words. This way,
we can describe commonalities of these guides using a section index
page, and make it easier to scale out these sections as we add related
guides:

- Key management service guides
- Automatic labeling

Also make minor page title changes so all entries in a sidebar have the
same grammatical structure.